+++ backend/src/main/java/com/kotonekanno/movielog/dto/ReviewDetailsDTO.java
@@ -3,7 +3,7 @@ package com.kotonekanno.movielog.dto;
 import java.time.LocalDate;
 
 public record ReviewDetailsDTO(
-    Long reviewId,
+    Integer reviewId,
     Double score,
     String text,
     LocalDate watchedAt,
